+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **🚀 New and Noteworthy:**
 
 **🐛 Fixed issues:**
+* [#71](https://github.com/pmd/pmd-designer/issues/71): Support JavaFX as module (e.g. via azul jdk+fx)
 
 **✨ Merged pull requests:**